vi-feat: moved HttpError exception into http package
