+++ client/src/libs/components/icon/common.js
@@ -13,7 +13,7 @@ import {
   faUser,
   faUserCircle
 } from '@fortawesome/free-solid-svg-icons';
-import { IconName } from 'libs/enums/enums';
+import { IconName } from 'libs/enums/enums.js';
 
 const iconNameToSvgIcon = {
   [IconName.AT]: faAt,
