+++ client/src/libs/components/post/post.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 
 import { getFromNowTime } from 'libs/helpers/helpers';
-import { IconName } from 'libs/enums/enums';
+import { IconName } from 'libs/enums/enums.js';
 import { postType } from 'libs/prop-types/prop-types';
 import { IconButton } from '../icon-button/icon-button.jsx';
 import { Image } from '../image/image.jsx';
