+++ client/src/packages/comment/comment-api.js
@@ -1,4 +1,4 @@
-import { ApiPath, CommentsApiPath, ContentType } from 'libs/enums/enums';
+import { ApiPath, CommentsApiPath, ContentType } from 'libs/enums/enums.js';
 import { HttpMethod } from 'packages/http/libs/enums/enums.js';
 
 class Comment {
