+++ client/src/packages/http/http-api.js
@@ -1,6 +1,6 @@
-import { HttpError } from 'libs/exceptions/exceptions';
 import { getStringifiedQuery } from 'libs/helpers/helpers';
-import { StorageKey } from 'libs/enums/enums';
+import { StorageKey } from 'libs/enums/enums.js';
+import { HttpError } from './libs/exceptions/exceptions.js';
 import { HttpHeader, HttpMethod } from './libs/enums/enums.js';
 
 class Http {
