+++ client/src/packages/http/libs/exceptions/exceptions.js
