+++ client/src/packages/image/image-api.js
@@ -1,4 +1,4 @@
-import { ApiPath, ImagePayloadKey } from 'libs/enums/enums';
+import { ApiPath, ImagePayloadKey } from 'libs/enums/enums.js';
 import { HttpMethod } from 'packages/http/libs/enums/enums.js';
 
 class Image {
