+++ client/src/packages/image/image.js
@@ -1,4 +1,4 @@
-import { ENV } from 'libs/enums/enums';
+import { ENV } from 'libs/enums/enums.js';
 import { http } from 'packages/http/http';
 import { Image } from './image-api.js';
