+++ client/src/pages/not-found/not-found.jsx
@@ -1,5 +1,5 @@
 import { NavLink } from 'react-router-dom';
-import { IconName, IconSize, AppRoute } from 'libs/enums/enums';
+import { IconName, IconSize, AppRoute } from 'libs/enums/enums.js';
 import { Icon } from 'libs/components/icon/icon.jsx';
 import styles from './styles.module.scss';
