+++ client/src/pages/shared-post/shared-post.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useDispatch, useParams } from 'libs/hooks/hooks';
 import { Navigate } from 'react-router-dom';
 import { actions as threadActionCreator } from 'slices/thread/thread.js';
-import { AppRoute } from 'libs/enums/enums';
+import { AppRoute } from 'libs/enums/enums.js';
 
 const SharedPost = () => {
   const dispatch = useDispatch();
