+++ client/src/pages/sign/components/login-form/libs/common/constants.js
@@ -1,4 +1,4 @@
-import { UserPayloadKey } from 'libs/enums/enums';
+import { UserPayloadKey } from 'libs/enums/enums.js';
 
 const DEFAULT_LOGIN_PAYLOAD = {
   [UserPayloadKey.EMAIL]: '',
